Version 1.62

+++ WEB-INF/classes/com/aoindustries/website/clientarea/accounting/DeleteCreditCardCompletedAction.java
@@ -1,10 +1,10 @@
+package com.aoindustries.website.clientarea.accounting;
+
 /*
- * Copyright 2007-2011 by AO Industries, Inc.,
+ * Copyright 2007-2009 by AO Industries, Inc.,
  * 7262 Bull Pen Cir, Mobile, Alabama, 36695, U.S.A.
  * All rights reserved.
  */
-package com.aoindustries.website.clientarea.accounting;
-
 import com.aoindustries.aoserv.client.AOServConnector;
 import com.aoindustries.aoserv.client.AOServPermission;
 import com.aoindustries.aoserv.client.CreditCard;
@@ -15,8 +15,10 @@ import com.aoindustries.aoserv.creditcards.CreditCardProcessorFactory;
 import com.aoindustries.website.PermissionAction;
 import com.aoindustries.website.SiteSettings;
 import com.aoindustries.website.Skin;
+import java.sql.SQLException;
 import java.util.Collections;
-import java.util.Set;
+import java.util.List;
+import java.util.Locale;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import org.apache.struts.action.ActionForm;
@@ -37,6 +39,7 @@ public class DeleteCreditCardCompletedAction extends PermissionAction {
         HttpServletRequest request,
         HttpServletResponse response,
         SiteSettings siteSettings,
+        Locale locale,
         Skin skin,
         AOServConnector aoConn
     ) throws Exception {
@@ -58,12 +61,13 @@ public class DeleteCreditCardCompletedAction extends PermissionAction {
         // Lookup the card in the root connector (to get access to the processor)
         AOServConnector rootConn = siteSettings.getRootAOServConnector();
         CreditCard rootCreditCard = rootConn.getCreditCards().get(creditCard.getPkey());
+        if(rootCreditCard==null) throw new SQLException("Unable to find CreditCard: "+creditCard.getPkey());
 
         // Delete the card from the bank and persistence
-        CreditCardProcessor rootAoservCCP = rootCreditCard.getProcessor();
+        CreditCardProcessor rootAoservCCP = rootCreditCard.getCreditCardProcessor();
         com.aoindustries.creditcards.CreditCardProcessor processor = CreditCardProcessorFactory.getCreditCardProcessor(rootAoservCCP);
         processor.deleteCreditCard(
-            new AOServConnectorPrincipal(rootConn, aoConn.getThisBusinessAdministrator().getUsername().getUsername().toString()),
+            new AOServConnectorPrincipal(rootConn, aoConn.getThisBusinessAdministrator().getUsername().getUsername()),
             CreditCardFactory.getCreditCard(rootCreditCard)
         );
 
@@ -74,8 +78,7 @@ public class DeleteCreditCardCompletedAction extends PermissionAction {
         return mapping.findForward("success");
     }
 
-    @Override
-    public Set<AOServPermission.Permission> getPermissions() {
-        return Collections.singleton(AOServPermission.Permission.delete_credit_card);
+    public List<AOServPermission.Permission> getPermissions() {
+        return Collections.singletonList(AOServPermission.Permission.delete_credit_card);
     }
 }
